#325: Created Endpoint for retrieving solutions

+++ src/main/java/org/benetech/mathshare/repository/ProblemSetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProblemSetRepository extends JpaRepository<ProblemSet, Long> {
+public interface ProblemSetRepository extends JpaRepository<ProblemSet, Integer> {
     ProblemSet findOneByEditCode(long editCode);
 }
